+++ containers/Connector/blockExplorer.ts
@@ -1,7 +1,6 @@
 import { OPTIMISM_NETWORKS } from '@synthetixio/optimism-networks';
 
-import { NetworkNameById, NetworkIdByName } from 'sdk/types/common';
-import { NetworkId } from 'sdk/types/common';
+import { NetworkId, NetworkNameById, NetworkIdByName } from 'sdk/types/common';
 
 export const getBaseUrl = (networkId: NetworkId) => {
 	if (networkId === 10 || networkId === 420) {
