--- sdk/constants/defaults.ts
@@ -1,1 +1,0 @@
-export const DEFAULT_NUMBER_OF_TRADES = 16;
