+++ sections/dashboard/SpotHistoryTable/SpotHistoryTable.tsx
@@ -15,10 +15,10 @@ import Connector from 'containers/Connector';
 import { blockExplorer } from 'containers/Connector/Connector';
 import useSelectedPriceCurrency from 'hooks/useSelectedPriceCurrency';
 import useGetWalletTrades from 'queries/synths/useGetWalletTrades';
+import { WalletTradesExchangeResult } from 'sdk/types/synths';
 import { useAppSelector } from 'state/hooks';
 import { selectSynthsMap } from 'state/wallet/selectors';
 import { ExternalLink } from 'styles/common';
-import { WalletTradesExchangeResult } from 'utils/exchange';
 
 import TimeDisplay from '../../futures/Trades/TimeDisplay';
