+++ state/app/actions.ts
@@ -5,6 +5,6 @@ import { ThunkConfig } from 'state/types';
 export const checkSynthetixStatus = createAsyncThunk<boolean, void, ThunkConfig>(
 	'app/checkSynthetixStatus',
 	(_, { extra: { sdk } }) => {
-		return sdk.general.getSynthetixStatus();
+		return sdk.system.getSynthetixStatus();
 	}
 );
